requirements.txt
aiogram==3.17.0
asyncpg==0.30.0
APScheduler==3.11.0
python-decouple
requests
redis==5.2.1
sqlalchemy==2.0.37
pydantic-settings==2.7.1
aiohttp
loguru==0.7.3
alembic==1.14.1